requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

